HWIDGEN removal script
This script can be used to remove HWIDGEN, in case you integrated it into your OS image. The original file had a different HWIDGEN extension name, but the developer decided (multiple times) to change the extension name e.g. from mk2 to mk3 and so on (for no reason). So this script (as it is now) searches for .mk6.